requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
motor==3.6.1
beanie==1.29.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.10.6
